Feat: 채팅 이어하기 기능 모듈화
