+++ front/src/components/IntroductionModal.js
@@ -85,4 +85,4 @@ const IntroductionModal = ({ isOpen, onClose, profileDetails }) => {
     );
 };
 
-export default IntroductionModal;
+export {IntroductionModal};
